requirements.txt
fastapi>=0.110,<0.116
uvicorn[standard]>=0.23,<0.30
pyjwt>=2.8,<3.0
pwdlib==0.3.0
python-multipart>=0.0.6,<0.1
python-dotenv>=1.0,<2.0
SQLAlchemy>=2.0,<2.1
pymysql>=1.1,<2.0